# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Inventario de Equipos Celulares",
    "version": "1.0.0",
    "description": "API REST para registrar, gestionar y monitorear el inventario de equipos celulares."
  },
  "servers": [
    {
      "url": "/",
      "description": "Servidor actual"
    }
  ],
  "paths": {
    "/devices": {
      "get": {
        "summary": "Lista equipos celulares",
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "IMEI, marca o modelo"
          },
          {
            "name": "brand",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceStatus"
            }
          },
          {
            "name": "locationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pagina de equipos"
          }
        }
      },
      "post": {
        "summary": "Registra un nuevo equipo",
        "tags": [
          "Devices"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Equipo creado"
          },
          "400": {
            "description": "Datos invalidos"
          },
          "409": {
            "description": "IMEI duplicado"
          }
        }
      }
    },
    "/devices/{id}": {
      "get": {
        "summary": "Detalle de un equipo",
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipo encontrado"
          },
          "404": {
            "description": "No existe"
          }
        }
      },
      "put": {
        "summary": "Actualiza un equipo",
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipo actualizado"
          },
          "404": {
            "description": "No existe"
          }
        }
      },
      "delete": {
        "summary": "Elimina un equipo",
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado"
          },
          "404": {
            "description": "No existe"
          }
        }
      }
    },
    "/movements": {
      "get": {
        "summary": "Historial de movimientos",
        "tags": [
          "Movements"
        ],
        "parameters": [
          {
            "name": "deviceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtra el historial de un equipo"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de movimientos"
          }
        }
      },
      "post": {
        "summary": "Registra un movimiento",
        "tags": [
          "Movements"
        ],
        "responses": {
          "201": {
            "description": "Movimiento creado"
          },
          "400": {
            "description": "Movimiento invalido"
          },
          "404": {
            "description": "El equipo no existe"
          }
        }
      }
    },
    "/reports/stock": {
      "get": {
        "summary": "Reporte de stock actual",
        "tags": [
          "Reports"
        ],
        "responses": {
          "200": {
            "description": "Reporte generado"
          }
        }
      }
    },
    "/locations": {
      "get": {
        "summary": "Lista las ubicaciones",
        "tags": [
          "Locations"
        ],
        "responses": {
          "200": {
            "description": "Ubicaciones"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Estado del servicio",
        "tags": [
          "Health"
        ],
        "responses": {
          "200": {
            "description": "Servicio disponible"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DeviceStatus": {
        "type": "string",
        "enum": [
          "Disponible",
          "Asignado",
          "EnMantenimiento",
          "DeBaja"
        ]
      },
      "MovementType": {
        "type": "string",
        "enum": [
          "Ingreso",
          "Salida",
          "Traslado",
          "Baja"
        ]
      },
      "CreateDeviceRequest": {
        "type": "object",
        "required": [
          "brand",
          "model",
          "imei",
          "locationId",
          "entryDate"
        ],
        "properties": {
          "brand": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "model": {
            "type": "string",
            "maxLength": 80
          },
          "imei": {
            "type": "string",
            "pattern": "^[0-9]{15}$",
            "description": "15 digitos con digito verificador valido (Luhn)"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "entryDate": {
            "type": "string",
            "format": "date-time"
          },
          "observations": {
            "type": "string",
            "maxLength": 500
          },
          "purchasePrice": {
            "type": "number",
            "minimum": 0
          }
        }
      }
    }
  }
}